pyproject.toml
[project]
name = "password-manager-latest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "amqp==5.2.0",
    "asgiref==3.8.1",
    "attrs==23.1.0",
    "billiard==4.2.0",
    "celery==5.3.6",
    "certifi==2023.11.17",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "click-didyoumean==0.3.0",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "deprecated==1.2.14",
    "django==5.0.6",
    "django-cors-headers==4.3.1",
    "django-filter==23.2",
    "django-hiken",
    "django-rest-passwordreset==1.3.0",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt==5.3.1",
    "drf-spectacular==0.26.4",
    "faker==22.1.0",
    "gunicorn==21.2.0",
    "idna==3.6",
    "importlib-resources==6.1.1",
    "inflection==0.5.1",
    "iniconfig==2.0.0",
    "jsonschema==4.18.6",
    "jsonschema-specifications==2023.7.1",
    "kombu==5.3.5",
    "limits==3.7.0",
    "markdown==3.5.1",
    "packaging==23.2",
    "pluggy==1.3.0",
    "prompt-toolkit==3.0.43",
    "psycopg2-binary==2.9.5",
    "pycryptodome==3.17",
    "pyjwt==2.6.0",
    "pylogbeat==2.0.1",
    "pytest==7.4.4",
    "pytest-django==4.7.0",
    "python-dateutil==2.8.2",
    "python-decouple==3.8",
    "python-dotenv==1.0.0",
    "python-logstash-async==2.7.2",
    "pytz==2022.7.1",
    "pyyaml==6.0.1",
    "redis==5.0.1",
    "referencing==0.30.2",
    "requests==2.31.0",
    "rosetta==0.3",
    "rpds-py==0.9.2",
    "six==1.16.0",
    "sqlparse==0.4.3",
    "typing-extensions==4.9.0",
    "tzdata==2023.4",
    "uritemplate==4.1.1",
    "urllib3==2.1.0",
    "vine==5.1.0",
    "wcwidth==0.2.13",
    "wrapt==1.16.0",
]

[tool.uv.sources]
django-hiken = { git = "https://github.com/jefri-noble/django_hiken.git" }
